+++ lib/string_splitter.rb
@@ -3,17 +3,39 @@
 require 'values'
 require_relative 'string_splitter/version'
 
+# terminology: the delimiter is what we provide and the separators are what we get
+# back (if we capture them). e.g. for:
+#
+#   ss.split("foo:bar::baz", /(\W+)/)
+#
+# the delimiter is /(\W)/ and the separators are ":" and "::"
+
 # This class extends the functionality of +String#split+ by:
 #
 #   - providing full control over which splits are accepted or rejected
+#
 #   - adding support for splitting from right-to-left
+#
 #   - encapsulating splitting options/preferences in the splitter rather
 #     than trying to cram them into overloaded method parameters
 #
 # These enhancements allow splits to handle many cases that otherwise require bigger
 # guns e.g. regex matching or parsing.
+#
+# Implementation-wise, we effectively use the built-in +String#split+ method as a
+# tokenizer, and parse the resulting tokens into an array of Split objects with the
+# following fields:
+#
+#   - captures:  separator substrings captured by parentheses in the delimiter pattern
+#   - count:     the number of splits
+#   - index:     the 0-based index of this split in the array
+#   - lhs:       the string to the left of the separator (back to the previous split candidate)
+#   - position:  the 1-based index of this split in the array
+#   - rhs:       the string to the right of the separator (up to the next split candidate)
+#   - separator: the string matched by the delimiter pattern/string
 class StringSplitter
   ACCEPT_ALL = ->(_split) { true }
+  DONE = Object.new
   DEFAULT_DELIMITER = /\s+/
   NO_SPLITS = []
 
@@ -37,27 +59,36 @@ class StringSplitter
   def initialize(
     default_delimiter: DEFAULT_DELIMITER,
     include_captures: true,
-    remove_empty: false,
+    remove_empty: false, # TODO remove this
+    remove_empty_fields: remove_empty,
     spread_captures: true
   )
     @default_delimiter = default_delimiter
     @include_captures = include_captures
-    @remove_empty = remove_empty
+    @remove_empty_fields = remove_empty_fields
     @spread_captures = spread_captures
   end
 
-  attr_reader :default_delimiter, :include_captures, :remove_empty, :spread_captures
+  attr_reader(
+    :default_delimiter,
+    :include_captures,
+    :remove_empty_fields,
+    :spread_captures
+  )
+
+  # TODO remove this
+  alias remove_empty remove_empty_fields
 
   def split(
     string,
     delimiter = @default_delimiter,
-    at: nil,
+    at: nil, # alias for select
+    exclude: nil, # alias for reject
     select: at,
-    exclude: nil,
     reject: exclude,
     &block
   )
-    result, splits, block = split_init(
+    result, splits, accept = init(
       string: string,
       delimiter: delimiter,
       select: select,
@@ -67,27 +98,19 @@ class StringSplitter
 
     count = splits.length
 
-    splits.each_with_index do |split, index|
-      split = Split.with(split.merge({ index: index, count: count }))
+    splits.each_with_index do |hash, index|
+      split = Split.with(hash.merge({ index: index, count: count }))
       result << split.lhs if result.empty?
 
-      if block.call(split)
-        if @include_captures
-          if @spread_captures
-            result += split.captures
-          else
-            result << split.captures
-          end
-        end
-
-        result << split.rhs
+      if accept.call(split)
+        result << split.captures << split.rhs
       else
         # concatenate the rhs
         result[-1] = result[-1] + split.separator + split.rhs
       end
     end
 
-    result
+    render(result)
   end
 
   alias lsplit split
@@ -95,13 +118,13 @@ class StringSplitter
   def rsplit(
     string,
     delimiter = @default_delimiter,
-    at: nil,
+    at: nil, # alias for select
+    exclude: nil, # alias for reject
     select: at,
-    exclude: nil,
     reject: exclude,
     &block
   )
-    result, splits, block = split_init(
+    result, splits, accept = init(
       string: string,
       delimiter: delimiter,
       select: select,
@@ -109,55 +132,77 @@ class StringSplitter
       block: block
     )
 
+    return result if accept == true
+
     count = splits.length
 
-    splits.reverse!.each_with_index do |split, index|
-      split = Split.with(split.merge({ index: index, count: count }))
+    splits.reverse!.each_with_index do |hash, index|
+      split = Split.with(hash.merge({ index: index, count: count }))
       result.unshift(split.rhs) if result.empty?
 
-      if block.call(split)
-        if @include_captures
-          if @spread_captures
-            result = split.captures + result
-          else
-            result.unshift(split.captures)
-          end
-        end
-
-        result.unshift(split.lhs)
+      if accept.call(split)
+        # [lhs + captures] + result
+        result.unshift(split.lhs, split.captures)
       else
         # prepend the lhs
         result[0] = split.lhs + split.separator + result[0]
       end
     end
 
-    result
+    render(result)
   end
 
   private
 
+  def render(result)
+    if @remove_empty_fields
+      result.reject! { |it| it.is_a?(String) && it.empty? }
+    end
+
+    if @include_captures
+      result.flat_map do |value|
+        next [value] unless value.is_a?(Array)
+        @spread_captures ? value : [value]
+      end
+    else
+      result.reject! { |it| it.is_a?(Array) }
+    end
+  end
+
+  # converts the tokens returned by
+  #
+  #   string.split(delimiter)
+  #
+  # into an array of objects (hashes) representing each split. e.g. for:
+  #
+  #   ss.split("foo:bar:baz:quux", ":")
+  #
+  # it's called with:
+  #
+  #   splits_for(["foo", ":", "bar", ":", "baz", ":", "quux"], 0)
+  #
+  # and returns:
+  #
+  #   [
+  #       { lhs: "foo", rhs: "bar", separator: ":", captures: [] },
+  #       { lhs: "bar", rhs: "baz", separator: ":", captures: [] },
+  #       { lhs: "baz", rhs: "quux", separator: ":", captures: [] },
+  #   ]
+
   def splits_for(parts, ncaptures)
-    result = []
     splits = []
+    rhs = nil
 
     until parts.empty?
-      lhs = parts.shift
+      lhs = rhs || parts.shift
       separator = parts.shift
       captures = parts.shift(ncaptures)
-      rhs = parts.length == 1 ? parts.shift : parts.first
+      rhs = parts.shift
 
-      if @remove_empty && (lhs.empty? || rhs.empty?)
-        if lhs.empty? && rhs.empty?
-          # do nothing
-        elsif parts.empty? # last split
-          result << (!lhs.empty? ? lhs : rhs) if splits.empty?
-        elsif rhs.empty?
-          # replace the empty rhs with the non-empty lhs
-          parts[0] = lhs
-        end
-
-        next
-      end
+      # there's no such thing as an empty field whose separator is empty, so if
+      # String#split's result ends with an empty separator, 0 or more (empty)
+      # captures and an empty field, we can safely remove them.
+      break if parts.empty? && separator.empty? && rhs.empty?
 
       splits << {
         lhs: lhs,
@@ -167,22 +212,36 @@ class StringSplitter
       }
     end
 
-    [result, splits]
+    splits
   end
 
-  # takes a hash of options passed to +split+ or +rsplit+ and returns a:
+  # takes a hash of options passed to +split+ or +rsplit+ and returns a triple with
+  # the following fields:
   #
-  #   [result, splits, block]
+  #   - result: the array of separated strings to return from +split+ or +rsplit+
+  #     if the accept value is DONE, the caller returns this result immediately
+  #     without any further processing
   #
-  # triple, where `result` is the return value of the method, `splits` is an array
-  # of hashes containing the lhs/rhs, separator and captures of each split, and
-  # `block` is a proc which specifies whether each split should be accepted or
-  # rejected
-  def split_init(string:, delimiter:, select:, reject:, block:)
-    unless (match = string.match(delimiter))
-      result = (@remove_empty && string.empty?) ? [] : [string]
-      return [result, NO_SPLITS, block]
+  #   - splits: an array of hashes containing the lhs, rhs, separator and captured
+  #     separator substrings for each split
+  #
+  #   - accept: a proc whose return value determines whether each split should be
+  #     accepted (true) or rejected (false)
+
+  def init(string:, delimiter:, select:, reject:, block:)
+    return [[], NO_SPLITS, DONE] if string.empty?
+
+    if delimiter.is_a?(String)
+      match = string.include?(delimiter)
+      ncaptures = 0
+    elsif (match = string.match(delimiter))
+      ncaptures = match.captures.length
     end
+
+    return [[string], NO_SPLITS, DONE] unless match
+
+    # XXX must be done *after* the include? test
+    delimiter = Regexp.quote(delimiter) if delimiter.is_a?(String)
 
     select = Array(select)
     reject = Array(reject)
@@ -195,15 +254,11 @@ class StringSplitter
       action = Action::SELECT
     end
 
-    ncaptures = match.captures.length
-    delimiter = Regexp.quote(delimiter) if delimiter.is_a?(String)
-    delimiter = increment_backrefs(delimiter, ncaptures)
-    parts = string.split(/(#{delimiter})/, -1)
-    remove_trailing_empty_field!(parts, ncaptures)
-    result, splits = splits_for(parts, ncaptures)
+    delimiter = increment_backrefs(/(#{delimiter})/)
+    parts = string.split(delimiter, -1)
+    splits = splits_for(parts, ncaptures)
     block ||= positions ? match_positions(positions, action, splits.length) : ACCEPT_ALL
-
-    [result, splits, block]
+    [[], splits, block]
   end
 
   # increment back-references so they remain valid when the outer capture
@@ -222,60 +277,18 @@ class StringSplitter
   #
   # after:
   #
+  #       +------- outer capture -------+
+  #       |                             |
+  #       v                             v
   #   %r| ( <(\w+-comment)> [^<]* </\2> ) |x
 
-  def increment_backrefs(delimiter, ncaptures)
-    if delimiter.is_a?(Regexp) && ncaptures > 0
-      delimiter = delimiter.to_s.gsub(/\\(?:(\d+)|.)/) do
-        match = Regexp.last_match
-        match[1] ? '\\' + match[1].to_i.next.to_s : match[0]
-      end
+  def increment_backrefs(delimiter)
+    delimiter = delimiter.to_s.gsub(/\\(?:(\d+)|.)/) do
+      match = Regexp.last_match
+      match[1] ? '\\' + match[1].to_i.next.to_s : match[0]
     end
 
-    delimiter
-  end
-
-  # work around Ruby's (and Perl's and Groovy's) unhelpful behavior when splitting
-  # on an empty string/pattern without removing trailing empty fields e.g.:
-  #
-  #   "foobar".split("", -1)
-  #   "foobar".split(//, -1)
-  #   # => ["f", "o", "o", "b", "a", "r", ""]
-  #
-  #   "foobar".split(/()/, -1)
-  #   # => ["f", "", "o", "", "o", "", "b", "", "a", "", "r", "", ""]
-  #
-  #   "foobar".split(/(())/, -1)
-  #   # => ["f", "", "", "o", "", "", "o", "", "", "b", "", "", "a", "", "", "r", "", "", ""]
-  #
-  # *there is no such thing as an empty field whose separator is empty*, so
-  # if String#split's result ends with an empty separator, 0 or more (empty)
-  # captures and an empty field, we can safely remove them.
-
-  def remove_trailing_empty_field!(parts, ncaptures)
-    # the trailing field is at index -1. if there are 0 captures, the separator
-    # is at -2:
-    #
-    #   [empty_separator, empty_field]
-    #
-    # if there is 1 capture, the separator is at -3:
-    #
-    #   [empty_separator, capture, empty_field]
-    #
-    # etc. therefore we find the separator by walking back
-    #
-    #  1 (empty field)
-    #  + ncaptures
-    #  + 1 (separator)
-    #
-    # steps from the end of the array i.e. ncaptures + 2
-    count = ncaptures + 2
-    separator_index = count * -1
-
-    return unless parts[-1].empty? && parts[separator_index].empty?
-
-    # drop the empty separator, the (empty) captures, and the trailing empty field
-    parts.pop(count)
+    Regexp.new(delimiter)
   end
 
   def match_positions(positions, action, nsplits)
@@ -290,11 +303,11 @@ class StringSplitter
         #   ss.split("foo:bar:baz:quux", ":", at: 3)
         #
         # XXX note: we don't use modulo, because we don't want
-        # out-of-bounds indices to silently work e.g. we don't want:
+        # out-of-bounds indices to silently work, e.g. we don't want:
         #
-        #   ss.split("foo:bar:baz:quux", ":", -42)
+        #   ss.split("foo:bar:baz:quux", ":", at: -42)
         #
-        # to mysteriously match when the position is 2
+        # to mysteriously match when the index/position is 0/1
 
         nsplits + 1 + position
       else