+++ test/limit.rb
@@ -4,27 +4,39 @@ require_relative 'test_helper'
 
 # confirm that String#split's limit parameter can be emulated
 
+string = 'foo:bar:baz:quux'
+ss = StringSplitter.new
+
 describe 'emulate limit' do
-  string = 'foo:bar:baz:quux'
-  s = StringSplitter.new
+  test 'limit: 0' do
+    result = ss.split(string, ':') { true }
+    assert { result == %w[foo bar baz quux] }
+
+    result = ss.rsplit(string, ':') { true }
+    assert { result == %w[foo bar baz quux] }
+  end
 
   test 'limit: 1' do
-    result = s.split(string, ':') { |split| split.pos < 1 }
+    result = ss.split(string, ':') { |split| split.pos < 1 }
+    assert { result == ['foo:bar:baz:quux'] }
+
+    result = ss.rsplit(string, ':') { |split| split.pos < 1 }
     assert { result == ['foo:bar:baz:quux'] }
   end
 
   test 'limit: 2' do
-    result = s.split(string, ':') { |split| split.pos < 2 }
+    result = ss.split(string, ':') { |split| split.pos < 2 }
     assert { result == %w[foo bar:baz:quux] }
+
+    result = ss.rsplit(string, ':') { |split| split.pos < 2 }
+    assert { result == %w[foo:bar:baz quux] }
   end
 
   test 'limit: 3' do
-    result = s.split(string, ':') { |split| split.pos < 3 }
+    result = ss.split(string, ':') { |split| split.pos < 3 }
     assert { result == %w[foo bar baz:quux] }
-  end
 
-  test 'limit: 0' do
-    result = s.split(string, ':') { true }
-    assert { result == %w[foo bar baz quux] }
+    result = ss.rsplit(string, ':') { |split| split.pos < 3 }
+    assert { result == %w[foo:bar baz quux] }
   end
 end