+++ test/reject.rb
@@ -6,29 +6,47 @@ require_relative 'test_helper'
 # indices
 
 describe 'reject' do
-  s = StringSplitter.new(default_delimiter: ':')
+  ss = StringSplitter.new(default_delimiter: ':')
 
   test 'reject a positive position' do
-    result = s.split('foo:bar:baz:quux', exclude: 1)
+    result = ss.split('foo:bar:baz:quux', exclude: 1)
     assert { result == %w[foo:bar baz quux] }
 
-    result = s.split('foo:bar:baz:quux', reject: 1)
+    result = ss.split('foo:bar:baz:quux', reject: 1)
     assert { result == %w[foo:bar baz quux] }
+
+    result = ss.rsplit('foo:bar:baz:quux', exclude: 1)
+    assert { result == %w[foo bar baz:quux] }
+
+    result = ss.rsplit('foo:bar:baz:quux', reject: 1)
+    assert { result == %w[foo bar baz:quux] }
   end
 
   test 'reject a negative position' do
-    result = s.split('foo:bar:baz:quux', exclude: -1)
+    result = ss.split('foo:bar:baz:quux', exclude: -1)
     assert { result == %w[foo bar baz:quux] }
 
-    result = s.split('foo:bar:baz:quux', reject: -1)
+    result = ss.split('foo:bar:baz:quux', reject: -1)
     assert { result == %w[foo bar baz:quux] }
+
+    result = ss.rsplit('foo:bar:baz:quux', exclude: -1)
+    assert { result == %w[foo:bar baz quux] }
+
+    result = ss.rsplit('foo:bar:baz:quux', reject: -1)
+    assert { result == %w[foo:bar baz quux] }
   end
 
   test 'reject multiple positions' do
-    result = s.split('foo:bar:baz:quux', exclude: [1, -1])
+    result = ss.split('foo:bar:baz:quux', exclude: [1, -1])
     assert { result == %w[foo:bar baz:quux] }
 
-    result = s.split('foo:bar:baz:quux', reject: [1, -1])
+    result = ss.split('foo:bar:baz:quux', reject: [1, -1])
+    assert { result == %w[foo:bar baz:quux] }
+
+    result = ss.rsplit('foo:bar:baz:quux', exclude: [1, -1])
+    assert { result == %w[foo:bar baz:quux] }
+
+    result = ss.rsplit('foo:bar:baz:quux', reject: [1, -1])
     assert { result == %w[foo:bar baz:quux] }
   end
 end