+++ test/remove_empty.rb
@@ -2,14 +2,14 @@
 
 require_relative 'test_helper'
 
-# confirm that the :remove_empty option removes empty fields
+# confirm that the :remove_empty_fields option removes empty fields
 
 describe 'remove_empty' do
-  s1 = StringSplitter.new # default: remove_empty: false
-  s2 = StringSplitter.new(remove_empty: false)
-  ss = StringSplitter.new(remove_empty: true)
+  s1 = StringSplitter.new # default: remove_empty_fields: false
+  s2 = StringSplitter.new(remove_empty_fields: false) # same as 1 but explicit
+  s3 = StringSplitter.new(remove_empty_fields: true)
 
-  it 'removes leading empty tokens' do
+  it 'removes leading empty fields' do
     string = ':foo:bar:baz:quux'
 
     result = s1.split(string, ':')
@@ -18,24 +18,42 @@ describe 'remove_empty' do
     result = s2.split(string, ':')
     assert { result == ['', 'foo', 'bar', 'baz', 'quux'] }
 
-    result = ss.split(string, ':') { true }
+    result = s3.split(string, ':') { true }
+    assert { result == %w[foo bar baz quux] }
+
+    result = s1.rsplit(string, ':')
+    assert { result == ['', 'foo', 'bar', 'baz', 'quux'] }
+
+    result = s2.rsplit(string, ':')
+    assert { result == ['', 'foo', 'bar', 'baz', 'quux'] }
+
+    result = s3.rsplit(string, ':') { true }
     assert { result == %w[foo bar baz quux] }
   end
 
-  it 'removes trailing empty tokens' do
+  it 'removes trailing empty fields' do
     string = 'foo:bar:baz:quux:'
 
-    result = s1.split(string, ':') { true }
+    result = s1.split(string, ':')
     assert { result == ['foo', 'bar', 'baz', 'quux', ''] }
 
-    result = s2.split(string, ':') { true }
+    result = s2.split(string, ':')
     assert { result == ['foo', 'bar', 'baz', 'quux', ''] }
 
-    result = ss.split(string, ':') { true }
+    result = s3.split(string, ':')
+    assert { result == %w[foo bar baz quux] }
+
+    result = s1.rsplit(string, ':')
+    assert { result == ['foo', 'bar', 'baz', 'quux', ''] }
+
+    result = s2.rsplit(string, ':')
+    assert { result == ['foo', 'bar', 'baz', 'quux', ''] }
+
+    result = s3.rsplit(string, ':')
     assert { result == %w[foo bar baz quux] }
   end
 
-  it 'removes embedded empty tokens' do
+  it 'removes embedded empty fields' do
     string = 'foo:bar::baz:quux'
 
     result = s1.split(string, ':')
@@ -44,11 +62,20 @@ describe 'remove_empty' do
     result = s2.split(string, ':')
     assert { result == ['foo', 'bar', '', 'baz', 'quux'] }
 
-    result = ss.split(string, ':')
+    result = s3.split(string, ':')
+    assert { result == %w[foo bar baz quux] }
+
+    result = s1.rsplit(string, ':')
+    assert { result == ['foo', 'bar', '', 'baz', 'quux'] }
+
+    result = s2.rsplit(string, ':')
+    assert { result == ['foo', 'bar', '', 'baz', 'quux'] }
+
+    result = s3.rsplit(string, ':')
     assert { result == %w[foo bar baz quux] }
   end
 
-  it 'removes all empty tokens' do
+  it 'removes all empty fields' do
     string = ':foo:bar::baz:quux:'
 
     result = s1.split(string, ':')
@@ -57,7 +84,16 @@ describe 'remove_empty' do
     result = s2.split(string, ':')
     assert { result == ['', 'foo', 'bar', '', 'baz', 'quux', ''] }
 
-    result = ss.split(string, ':')
+    result = s3.split(string, ':')
+    assert { result == %w[foo bar baz quux] }
+
+    result = s1.rsplit(string, ':')
+    assert { result == ['', 'foo', 'bar', '', 'baz', 'quux', ''] }
+
+    result = s2.rsplit(string, ':')
+    assert { result == ['', 'foo', 'bar', '', 'baz', 'quux', ''] }
+
+    result = s3.rsplit(string, ':')
     assert { result == %w[foo bar baz quux] }
   end
 
@@ -70,7 +106,16 @@ describe 'remove_empty' do
     result = s2.split(string, ':')
     assert { result == ['', '', '', '', ''] }
 
-    result = ss.split(string, ':')
+    result = s3.split(string, ':')
+    assert { result == [] }
+
+    result = s1.rsplit(string, ':')
+    assert { result == ['', '', '', '', ''] }
+
+    result = s2.rsplit(string, ':')
+    assert { result == ['', '', '', '', ''] }
+
+    result = s3.rsplit(string, ':')
     assert { result == [] }
   end
 
@@ -83,7 +128,16 @@ describe 'remove_empty' do
     result = s2.split(string, ':')
     assert { result == ['', '', '', 'foo', '', '', ''] }
 
-    result = ss.split(string, ':')
+    result = s3.split(string, ':')
+    assert { result == ['foo'] }
+
+    result = s1.rsplit(string, ':')
+    assert { result == ['', '', '', 'foo', '', '', ''] }
+
+    result = s2.rsplit(string, ':')
+    assert { result == ['', '', '', 'foo', '', '', ''] }
+
+    result = s3.rsplit(string, ':')
     assert { result == ['foo'] }
   end
 end