+++ test/select.rb
@@ -6,29 +6,47 @@ require_relative 'test_helper'
 # indices
 
 describe 'select' do
-  s = StringSplitter.new(default_delimiter: ':')
+  ss = StringSplitter.new(default_delimiter: ':')
 
   test 'select a positive position' do
-    result = s.split('foo:bar:baz:quux', at: 1)
+    result = ss.split('foo:bar:baz:quux', at: 1)
     assert { result == %w[foo bar:baz:quux] }
 
-    result = s.split('foo:bar:baz:quux', select: 1)
+    result = ss.split('foo:bar:baz:quux', select: 1)
     assert { result == %w[foo bar:baz:quux] }
+
+    result = ss.rsplit('foo:bar:baz:quux', at: 1)
+    assert { result == %w[foo:bar:baz quux] }
+
+    result = ss.rsplit('foo:bar:baz:quux', select: 1)
+    assert { result == %w[foo:bar:baz quux] }
   end
 
   test 'select a negative position' do
-    result = s.split('foo:bar:baz:quux', at: -1)
+    result = ss.split('foo:bar:baz:quux', at: -1)
     assert { result == %w[foo:bar:baz quux] }
 
-    result = s.split('foo:bar:baz:quux', select: -1)
+    result = ss.split('foo:bar:baz:quux', select: -1)
     assert { result == %w[foo:bar:baz quux] }
+
+    result = ss.rsplit('foo:bar:baz:quux', at: -1)
+    assert { result == %w[foo bar:baz:quux] }
+
+    result = ss.rsplit('foo:bar:baz:quux', select: -1)
+    assert { result == %w[foo bar:baz:quux] }
   end
 
   test 'select multiple positions' do
-    result = s.split('foo:bar:baz:quux', at: [1, -1])
+    result = ss.split('foo:bar:baz:quux', at: [1, -1])
     assert { result == %w[foo bar:baz quux] }
 
-    result = s.split('foo:bar:baz:quux', select: [1, -1])
+    result = ss.split('foo:bar:baz:quux', select: [1, -1])
+    assert { result == %w[foo bar:baz quux] }
+
+    result = ss.rsplit('foo:bar:baz:quux', at: [1, -1])
+    assert { result == %w[foo bar:baz quux] }
+
+    result = ss.rsplit('foo:bar:baz:quux', select: [1, -1])
     assert { result == %w[foo bar:baz quux] }
   end
 end