+++ test/string_delimiter.rb
@@ -2,28 +2,40 @@
 
 require_relative 'test_helper'
 
-# make sure string delimiters are matched as strings rather than patterns
+# make sure string delimiters are matched verbatim rather than as patterns
 
 describe 'string delimiter' do
-  s = StringSplitter.new
+  s1 = StringSplitter.new
 
   it 'handles non-special delimiters' do
-    ss = StringSplitter.new(default_delimiter: ':')
+    s2 = StringSplitter.new(default_delimiter: ':')
 
-    result = s.split('foo:bar:baz', ':')
+    result = s1.split('foo:bar:baz', ':')
     assert { result == %w[foo bar baz] }
 
-    result = ss.split('foo:bar:baz')
+    result = s2.split('foo:bar:baz')
+    assert { result == %w[foo bar baz] }
+
+    result = s1.rsplit('foo:bar:baz', ':')
+    assert { result == %w[foo bar baz] }
+
+    result = s2.rsplit('foo:bar:baz')
     assert { result == %w[foo bar baz] }
   end
 
   it 'handles special delimiters' do
-    ss = StringSplitter.new(default_delimiter: '|')
+    s2 = StringSplitter.new(default_delimiter: '|')
 
-    result = s.split('foo|bar|baz', '|')
+    result = s1.split('foo|bar|baz', '|')
     assert { result == %w[foo bar baz] }
 
-    result = ss.split('foo|bar|baz')
+    result = s2.split('foo|bar|baz')
+    assert { result == %w[foo bar baz] }
+
+    result = s1.rsplit('foo|bar|baz', '|')
+    assert { result == %w[foo bar baz] }
+
+    result = s2.rsplit('foo|bar|baz')
     assert { result == %w[foo bar baz] }
   end
 end